IdeaForge_Backend/requirements.txt
Flask
python-dotenv
requests
google-cloud-storage
google-cloud-build
google-auth 